requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
clickhouse-connect==0.6.13
clickhouse-driver
clickhouse-sqlalchemy
aioch
httpx
requests
websockets
paramiko==3.4.0
pandas
plotly
streamlit
apscheduler
tqdm
